+++ plugins/communicators/ssh/communicator.rb
@@ -57,8 +57,18 @@ module VagrantPlugins
         }.merge(opts || {})
 
         # Connect via SSH and execute the command in the shell.
+        stdout = ""
+        stderr = ""
         exit_status = connect do |connection|
-          shell_execute(connection, command, opts[:sudo], &block)
+          shell_execute(connection, command, opts[:sudo]) do |type, data|
+            if type == :stdout
+              stdout += data
+            elsif type == :stderr
+              stderr += data
+            end
+
+            block.call(type, data) if block
+          end
         end
 
         # Check for any errors
@@ -66,7 +76,11 @@ module VagrantPlugins
           # The error classes expect the translation key to be _key,
           # but that makes for an ugly configuration parameter, so we
           # set it here from `error_key`
-          error_opts = opts.merge(:_key => opts[:error_key])
+          error_opts = opts.merge(
+            :_key => opts[:error_key],
+            :stdout => stdout,
+            :stderr => stderr
+          )
           raise opts[:error_class], error_opts
         end